Add sass gulp task
The first thing it does is find all .scss files that
it can see in eleventy-from-scratch/src/scss. These
are our main Sass files. We then get those files and
pass them down the production line into another
pipe. This is where the sassProcessor grabs them and
runs them through sass.

Once Sass is done, we pass it down the production
line to cleanCSS. This will compress, merge and
minify CSS based on whether isProduction is true or
false. The level: 2 essentially tells it to use a
lot of the most aggressive rules.

(opens new window).

Lastly, we pass it to our calculateOutput which will determine if it is critical CSS or not.